use routes rather than gulp
