+++ Jenkinsfile_CNP
@@ -41,9 +41,6 @@ withPipeline("nodejs", product, component) {
     after('checkout') {yarnBuilder.yarn('cache clean')}
     after('build') {
         yarnBuilder.yarn('build')
-        stage('Generate cookie banner') {
-            sh "./node_modules/gulp/bin/gulp.js default"
-        }
     }
 
     enableAksStagingDeployment()
